# pkg/query: Improve filter performance