1) Only lowercase and convert filter string to bytes once.
2) Run filter on dictionary only and then match indices. 
3) Run filter on all lines of a sample's locations at once to exit loop
   earlier.

Filtering itself is 52% faster.

```
$ benchstat old.txt new.txt
name                  old time/op  new time/op  delta
FilterProfileData-10  0.05ns ± 1%  0.02ns ± 4%  -52.41%  (p=0.000 n=10+10)
```

And requests in total that filter about 17%.

```
$ benchstat old.txt new.txt
name                            old time/op  new time/op  delta
ColumnQueryAPIQueryFiltered-10   136ms ± 3%   112ms ± 4%  -17.62%  (p=0.008 n=5+5)
```